Implementa layout CoberturaRiscoSeguro

+++ Models/Beneficiario.cs
@@ -5,7 +5,7 @@ public class Beneficiario
     public long Id { get; set; }
     public long DocumentoId { get; set; }
     public string DocumentoIdentificacao { get; set; } = string.Empty;
-    public short TipoDocumento { get; set; }
+    public int TipoDocumento { get; set; }
     public string Nome { get; set; } = string.Empty;
     public string? NomeSocial { get; set; }
     public string CodigoPostal { get; set; } = string.Empty;
